Move search to bottom panel and make 'esc' close everything
When doing search in files, the output is too hard to read in the side bar. Move it to the bottom pabel which had more width.

Use macros to map the escape key to close the bottom panel and any open find widget. Need to use this macros thing because keybindings currently don't allow one key to do multiple things.